+++ Source/DotNetWorkQueue.Transport.Redis/Basic/CommandHandler/MoveRecordToErrorQueueCommandHandlerAsync.cs
@@ -1,0 +1,52 @@
+﻿// ---------------------------------------------------------------------
+//This file is part of DotNetWorkQueue
+//Copyright © 2015-2026 Brian Lehnen
+//
+//This library is free software; you can redistribute it and/or
+//modify it under the terms of the GNU Lesser General Public
+//License as published by the Free Software Foundation; either
+//version 2.1 of the License, or (at your option) any later version.
+//
+//This library is distributed in the hope that it will be useful,
+//but WITHOUT ANY WARRANTY; without even the implied warranty of
+//MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+//Lesser General Public License for more details.
+//
+//You should have received a copy of the GNU Lesser General Public
+//License along with this library; if not, write to the Free Software
+//Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
+// ---------------------------------------------------------------------
+using System.Threading.Tasks;
+using DotNetWorkQueue.Transport.Redis.Basic.Command;
+using DotNetWorkQueue.Transport.Redis.Basic.Lua;
+using DotNetWorkQueue.Transport.Shared;
+using DotNetWorkQueue.Transport.Shared.Basic.Command;
+using DotNetWorkQueue.Validation;
+
+namespace DotNetWorkQueue.Transport.Redis.Basic.CommandHandler
+{
+    /// <inheritdoc />
+    internal class MoveRecordToErrorQueueCommandHandlerAsync : ICommandHandlerAsync<MoveRecordToErrorQueueCommand<string>>
+    {
+        private readonly ErrorLua _errorLua;
+        private readonly IUnixTime _unixTime;
+
+        /// <summary>Initializes a new instance of the <see cref="DeleteMessageCommandHandler"/> class.</summary>
+        /// <param name="errorLua">The error lua.</param>
+        /// <param name="timeFactory">Time factory</param>
+        public MoveRecordToErrorQueueCommandHandlerAsync(ErrorLua errorLua, IUnixTimeFactory timeFactory)
+        {
+            Guard.NotNull(errorLua);
+            Guard.NotNull(timeFactory);
+            _errorLua = errorLua;
+            _unixTime = timeFactory.Create();
+        }
+
+        /// <inheritdoc />
+        public async Task HandleAsync(MoveRecordToErrorQueueCommand<string> command)
+        {
+            await _errorLua.ExecuteAsync(command.QueueId, _unixTime.GetCurrentUnixTimestampMilliseconds())
+                .ConfigureAwait(false);
+        }
+    }
+}